tests: Move `helpers` folder out of `unit`
