+++ tests/helpers/format-num-test.js
@@ -1,7 +1,7 @@
-import { formatNum } from '../../../helpers/format-num';
+import { formatNum } from '../../helpers/format-num';
 import { module, test } from 'qunit';
 
-module('Unit | Helper | format-num', function() {
+module('Helper | format-num', function() {
   test('it works', function(assert) {
     assert.equal(formatNum(42), '42');
     assert.equal(formatNum(0), '0');
